+++ core/api_container/server/service_network/service_network.go
@@ -40,6 +40,13 @@ type ServiceNetwork interface {
 		filesToGenerate map[string]*core_api_bindings.FileGenerationOptions,
 	) (map[string]string, error)
 
+	// Copies files from the static file cache to the given service's filespace
+	// Returns a mapping of static_file_id -> filepath_relative_to_suite_ex_vol_root
+	LoadStaticFiles(
+		serviceId service_network_types.ServiceID,
+		staticFileIdKeys map[string]bool,
+	) (map[string]string, error)
+
 	// TODO add tests for this
 	/*
 	Starts a previously-registered but not-started service by creating it in a container
@@ -74,8 +81,5 @@ type ServiceNetwork interface {
 
 	GetServiceIP(serviceId service_network_types.ServiceID) (net.IP, error)
 
-	Destroy(
-		ctx context.Context,
-		containerStopTimeout time.Duration,
-	) error
+	GetServiceSuiteExecutionVolMntDirpath(serviceId service_network_types.ServiceID) (string, error)
 }