+++ core/api_container/server/service_network/service_network_impl.go
@@ -18,7 +18,9 @@ import (
 	"github.com/kurtosis-tech/kurtosis/commons/suite_execution_volume"
 	"github.com/palantir/stacktrace"
 	"github.com/sirupsen/logrus"
+	"io"
 	"net"
+	"os"
 	"strings"
 	"sync"
 	"time"
@@ -28,6 +30,21 @@ const (
 	defaultPartitionId                   service_network_types.PartitionID = "default"
 	startingDefaultConnectionBlockStatus                                   = false
 )
+
+// Information that gets created with a service's registration
+type serviceRegistrationInfo struct {
+	ipAddr net.IP
+	serviceDirectory *suite_execution_volume.ServiceDirectory
+}
+
+// Information that gets created when a container is started for a service
+type serviceRunInfo struct {
+	// Service's container ID
+	containerId string
+
+	// Where the suite execution volume is mounted on the service
+	suiteExecutionVolumeMountDirpath string
+}
 
 /*
 This is the in-memory representation of the service network that the API container will manipulate. To make
@@ -50,14 +67,16 @@ type ServiceNetworkImpl struct {
 
 	testExecutionDirectory *suite_execution_volume.EnclaveDirectory
 
+	staticFileCache *suite_execution_volume.StaticFileCache
+
 	userServiceLauncher *user_service_launcher.UserServiceLauncher
 
 	topology *partition_topology.PartitionTopology
 
 	// These are separate maps, rather than being bundled into a single containerInfo-valued map, because
 	//  they're registered at different times (rather than in one atomic operation)
-	serviceIps map[service_network_types.ServiceID]net.IP
-	serviceContainerIds map[service_network_types.ServiceID]string
+	serviceRegistrationInfo map[service_network_types.ServiceID]serviceRegistrationInfo
+	serviceRunInfo map[service_network_types.ServiceID]serviceRunInfo
 
 	networkingSidecars map[service_network_types.ServiceID]networking_sidecar.NetworkingSidecar
 
@@ -69,6 +88,7 @@ func NewServiceNetworkImpl(
 		freeIpAddrTracker *commons.FreeIpAddrTracker,
 		dockerManager *docker_manager.DockerManager,
 		testExecutionDirectory *suite_execution_volume.EnclaveDirectory,
+		staticFileCache *suite_execution_volume.StaticFileCache,
 		userServiceLauncher *user_service_launcher.UserServiceLauncher,
 		networkingSidecarManager networking_sidecar.NetworkingSidecarManager) *ServiceNetworkImpl {
 	defaultPartitionConnection := partition_topology.PartitionConnection{IsBlocked: startingDefaultConnectionBlockStatus}
@@ -78,14 +98,15 @@ func NewServiceNetworkImpl(
 		freeIpAddrTracker: freeIpAddrTracker,
 		dockerManager: dockerManager,
 		testExecutionDirectory: testExecutionDirectory,
+		staticFileCache: staticFileCache,
 		userServiceLauncher: userServiceLauncher,
 		mutex:               &sync.Mutex{},
 		topology:            partition_topology.NewPartitionTopology(
 			defaultPartitionId,
 			defaultPartitionConnection,
 		),
-		serviceIps:               map[service_network_types.ServiceID]net.IP{},
-		serviceContainerIds:      map[service_network_types.ServiceID]string{},
+		serviceRegistrationInfo:  map[service_network_types.ServiceID]serviceRegistrationInfo{},
+		serviceRunInfo:           map[service_network_types.ServiceID]serviceRunInfo{},
 		networkingSidecars:       map[service_network_types.ServiceID]networking_sidecar.NetworkingSidecar{},
 		networkingSidecarManager: networkingSidecarManager,
 	}
@@ -117,7 +138,7 @@ func (network *ServiceNetworkImpl) Repartition(
 		return stacktrace.Propagate(err, "An error occurred getting the blocklists after repartition, meaning that " +
 			"no partitions are actually being enforced!")
 	}
-	if err := updateIpTables(ctx, blocklists, network.serviceIps, network.networkingSidecars); err != nil {
+	if err := updateIpTables(ctx, blocklists, network.serviceRegistrationInfo, network.networkingSidecars); err != nil {
 		return stacktrace.Propagate(err, "An error occurred updating the IP tables to match the target blocklist after repartitioning")
 	}
 	return nil
@@ -139,8 +160,7 @@ func (network ServiceNetworkImpl) RegisterService(
 		return nil, stacktrace.NewError("Service ID cannot be empty or whitespace")
 	}
 
-	_, found := network.serviceIps[serviceId]
-	if found {
+	if _, found := network.serviceRegistrationInfo[serviceId]; found {
 		return nil, stacktrace.NewError("Cannot register service with ID '%v'; a service with that ID already exists", serviceId)
 	}
 
@@ -154,18 +174,33 @@ func (network ServiceNetworkImpl) RegisterService(
 		)
 	}
 
+	serviceDirectory, err := network.testExecutionDirectory.NewServiceDirectory(string(serviceId))
+	if err != nil {
+		return nil, stacktrace.Propagate(err, "An error occurred creating a new service directory for service with ID '%v'", serviceId)
+	}
 	ip, err := network.freeIpAddrTracker.GetFreeIpAddr()
 	if err != nil {
 		return nil, stacktrace.Propagate(err, "An error occurred getting an IP for service with ID '%v'", serviceId)
 	}
-	logrus.Debugf("Giving service '%v' IP '%v'", serviceId, ip.String())
-	network.serviceIps[serviceId] = ip
-	shouldUndoServiceIpAdd := true
-	// To keep our bookkeeping correct, if an error occurs later we need to back out the IP-adding that we do now
+	shouldFreeIpAddr := true
 	defer func() {
-		if shouldUndoServiceIpAdd {
-			delete(network.serviceIps, serviceId)
+		// To keep our bookkeeping correct, if an error occurs later we need to back out the IP-adding that we do now
+		if shouldFreeIpAddr {
 			network.freeIpAddrTracker.ReleaseIpAddr(ip)
+		}
+	}()
+	logrus.Debugf("Giving service '%v' IP '%v'", serviceId, ip.String())
+
+	registrationInfo := serviceRegistrationInfo{
+		ipAddr:           ip,
+		serviceDirectory: serviceDirectory,
+	}
+	network.serviceRegistrationInfo[serviceId] = registrationInfo
+	shouldUndoRegistrationInfoAdd := true
+	defer func() {
+		// If an error occurs, the service ID won't be used so we need to delete it from the map
+		if shouldUndoRegistrationInfoAdd {
+			delete(network.serviceRegistrationInfo, serviceId)
 		}
 	}()
 
@@ -176,8 +211,9 @@ func (network ServiceNetworkImpl) RegisterService(
 			serviceId,
 			partitionId)
 	}
-	shouldUndoServiceIpAdd = false
 
+	shouldFreeIpAddr = false
+	shouldUndoRegistrationInfoAdd = false
 	return ip, nil
 }
 
@@ -192,17 +228,18 @@ func (network *ServiceNetworkImpl) GenerateFiles(
 		return nil, stacktrace.NewError("Cannot register service with ID '%v'; the service network has been destroyed", serviceId)
 	}
 
-	serviceDirectory, err := network.testExecutionDirectory.GetServiceDirectory(string(serviceId))
-	if err != nil {
-		return nil, stacktrace.Propagate(err, "An error occurred getting the service directory inside the test execution volume for service with ID '%v'", serviceId)
+	registrationInfo, found := network.serviceRegistrationInfo[serviceId]
+	if !found {
+		return nil, stacktrace.NewError("Cannot generate files for service '%v' as no service with that ID has been registered", serviceId)
 	}
+	serviceDirectory := registrationInfo.serviceDirectory
 
 	generatedFilesRelativeFilepaths := map[string]string{}
 	for userCreatedFileKey, fileGenerationOptions := range filesToGenerate {
 		fileTypeToGenerate := fileGenerationOptions.GetFileTypeToGenerate()
 		switch fileTypeToGenerate {
 		case core_api_bindings.FileGenerationOptions_FILE:
-			file, err := serviceDirectory.GetFile(userCreatedFileKey)
+			file, err := serviceDirectory.NewGeneratedFile(userCreatedFileKey)
 			if err != nil {
 				return nil, stacktrace.Propagate(err, "An error occurred creating file '%v' for service with ID '%v'", userCreatedFileKey, serviceId)
 			}
@@ -217,6 +254,55 @@ func (network *ServiceNetworkImpl) GenerateFiles(
 		}
 	}
 	return generatedFilesRelativeFilepaths, nil
+}
+
+func (network *ServiceNetworkImpl) LoadStaticFiles(serviceId service_network_types.ServiceID, staticFileIdKeys map[string]bool) (map[string]string, error) {
+	// TODO extract this into a wrapper function that can be wrapped around every service call (so we don't forget)
+	network.mutex.Lock()
+	defer network.mutex.Unlock()
+	if network.isDestroyed {
+		return nil, stacktrace.NewError("Cannot load static files for service with ID '%v'; the service network has been destroyed", serviceId)
+	}
+
+	registrationInfo, found := network.serviceRegistrationInfo[serviceId]
+	if !found {
+		return nil, stacktrace.NewError("Cannot load static files for service '%v' as no service with that ID has been registered", serviceId)
+	}
+	serviceDirectory := registrationInfo.serviceDirectory
+
+	result := map[string]string{}
+	for staticFileId := range staticFileIdKeys {
+		src, err := network.staticFileCache.GetEntry(staticFileId)
+		if err != nil {
+			return nil, stacktrace.Propagate(err, "An error occurred getting the source static file with key '%v' from the static file cache", staticFileId)
+		}
+
+		dest, err := serviceDirectory.NewStaticFile(staticFileId)
+		if err != nil {
+			return nil, stacktrace.Propagate(err, "An error occurred getting the destination static file for key '%v' in filespace of service '%v'", staticFileId, serviceId)
+		}
+
+		srcAbsFilepath := src.GetAbsoluteFilepath()
+		srcFp, err := os.Open(srcAbsFilepath)
+		if err != nil {
+			return nil, stacktrace.Propagate(err, "An error occurred getting the file pointer from source static file filepath '%v'", srcAbsFilepath)
+		}
+		defer srcFp.Close()
+
+		destAbsFilepath := dest.GetAbsoluteFilepath()
+		destFp, err := os.Create(destAbsFilepath)
+		if err != nil {
+			return nil, stacktrace.Propagate(err, "An error occurred creating the file pointer for destination static file filepath '%v'", destAbsFilepath)
+		}
+		defer destFp.Close()
+
+		if _, err := io.Copy(destFp, srcFp); err != nil {
+			return nil, stacktrace.Propagate(err, "An error occurred copying source static file '%v' to destination '%v'", srcAbsFilepath, destAbsFilepath)
+		}
+
+		result[staticFileId] = src.GetFilepathRelativeToVolRoot()
+	}
+	return result, nil
 }
 
 // TODO add tests for this
@@ -244,12 +330,12 @@ func (network *ServiceNetworkImpl) StartService(
 		return nil, stacktrace.NewError("Cannot start container for service with ID '%v'; the service network has been destroyed", serviceId)
 	}
 
-	serviceIpAddr, foundIp := network.serviceIps[serviceId]
-	if !foundIp {
+	registrationInfo, registrationInfoFound := network.serviceRegistrationInfo[serviceId]
+	if !registrationInfoFound {
 		return nil, stacktrace.NewError("Cannot start container for service with ID '%v'; no service with that ID has been registered", serviceId)
 	}
-	if _, found := network.serviceContainerIds[serviceId]; found {
-		return nil, stacktrace.NewError("Cannot start container for service with ID '%v'; that service ID already has a container associated with it", serviceId)
+	if _, found := network.serviceRunInfo[serviceId]; found {
+		return nil, stacktrace.NewError("Cannot start container for service with ID '%v'; that service ID already has run information associated with it", serviceId)
 	}
 
 	// When partitioning is enabled, there's a race condition where:
@@ -272,7 +358,7 @@ func (network *ServiceNetworkImpl) StartService(
 			}
 			blocklistsWithoutNewNode[serviceInTopologyId] = servicesToBlock
 		}
-		if err := updateIpTables(ctx, blocklistsWithoutNewNode, network.serviceIps, network.networkingSidecars); err != nil {
+		if err := updateIpTables(ctx, blocklistsWithoutNewNode, network.serviceRegistrationInfo, network.networkingSidecars); err != nil {
 			return nil, stacktrace.Propagate(err, "An error occurred updating the iptables of all the other services " +
 				"before adding the node, meaning that the node wouldn't actually start in a partition")
 		}
@@ -281,7 +367,7 @@ func (network *ServiceNetworkImpl) StartService(
 	serviceContainerId, hostPortBindings, err := network.userServiceLauncher.Launch(
 		ctx,
 		serviceId,
-		serviceIpAddr,
+		registrationInfo.ipAddr,
 		imageName,
 		usedPorts,
 		entrypointArgs,
@@ -294,7 +380,11 @@ func (network *ServiceNetworkImpl) StartService(
 			err,
 			"An error occurred creating the user service container")
 	}
-	network.serviceContainerIds[serviceId] = serviceContainerId
+	runInfo := serviceRunInfo{
+		containerId:                      serviceContainerId,
+		suiteExecutionVolumeMountDirpath: suiteExecutionVolMntDirpath,
+	}
+	network.serviceRunInfo[serviceId] = runInfo
 
 	if network.isPartitioningEnabled {
 		sidecar, err := network.networkingSidecarManager.Add(ctx, serviceId, serviceContainerId)
@@ -318,7 +408,7 @@ func (network *ServiceNetworkImpl) StartService(
 		updatesToApply := map[service_network_types.ServiceID]*service_network_types.ServiceIDSet{
 			serviceId: newNodeBlocklist,
 		}
-		if err := updateIpTables(ctx, updatesToApply, network.serviceIps, network.networkingSidecars); err != nil {
+		if err := updateIpTables(ctx, updatesToApply, network.serviceRegistrationInfo, network.networkingSidecars); err != nil {
 			return nil, stacktrace.Propagate(err, "An error occurred applying the iptables on the new node to partition it " +
 				"off from other nodes")
 		}
@@ -356,7 +446,7 @@ func (network *ServiceNetworkImpl) ExecCommand(
 		return 0, nil, stacktrace.NewError("Cannot run exec command; the service network has been destroyed")
 	}
 
-	containerId, found := network.serviceContainerIds[serviceId]
+	runInfo, found := network.serviceRunInfo[serviceId]
 	if !found {
 		return 0, nil, stacktrace.NewError(
 			"Could not run exec command '%v' against service '%v'; no container has been created for the service yet",
@@ -368,7 +458,7 @@ func (network *ServiceNetworkImpl) ExecCommand(
 	// In the future, this will likely be insufficient
 
 	execOutputBuf := &bytes.Buffer{}
-	exitCode, err := network.dockerManager.RunExecCommand(ctx, containerId, command, execOutputBuf)
+	exitCode, err := network.dockerManager.RunExecCommand(ctx, runInfo.containerId, command, execOutputBuf)
 	if err != nil {
 		return 0, nil, stacktrace.Propagate(
 			err,
@@ -379,18 +469,34 @@ func (network *ServiceNetworkImpl) ExecCommand(
 	return exitCode, execOutputBuf, nil
 }
 
-func (network *ServiceNetworkImpl)  GetServiceIP(serviceId service_network_types.ServiceID,) (net.IP, error) {
+func (network *ServiceNetworkImpl) GetServiceIP(serviceId service_network_types.ServiceID) (net.IP, error) {
 	network.mutex.Lock()
 	defer network.mutex.Unlock()
 	if network.isDestroyed {
 		return nil, stacktrace.NewError("Cannot get IP address; the service network has been destroyed")
 	}
 
-	ip, found := network.serviceIps[serviceId]; if !found {
+	registrationInfo, found := network.serviceRegistrationInfo[serviceId]
+	if !found {
 		return nil, stacktrace.NewError("Service with ID: '%v' does not exist", serviceId)
 	}
 
-	return ip, nil
+	return registrationInfo.ipAddr, nil
+}
+
+func (network *ServiceNetworkImpl) GetServiceSuiteExecutionVolMntDirpath(serviceId service_network_types.ServiceID) (string, error) {
+	network.mutex.Lock()
+	defer network.mutex.Unlock()
+	if network.isDestroyed {
+		return "", stacktrace.NewError("Cannot get suite execution volume mount directory path; the service network has been destroyed")
+	}
+
+	runInfo, found := network.serviceRunInfo[serviceId]
+	if !found {
+		return "", stacktrace.NewError("No run information found for service with ID '%v'", serviceId)
+	}
+
+	return runInfo.suiteExecutionVolumeMountDirpath, nil
 }
 
 // Destroy all services the network is tracking, as well as make the network not usable anymore
@@ -432,25 +538,27 @@ func (network *ServiceNetworkImpl) removeServiceWithoutMutex(
 		ctx context.Context,
 		serviceId service_network_types.ServiceID,
 		containerStopTimeout time.Duration) error {
-	serviceIp, foundIpAddr := network.serviceIps[serviceId]
-	if !foundIpAddr {
-		return stacktrace.NewError("No IP found for service '%v'", serviceId)
+	registrationInfo, foundRegistrationInfo := network.serviceRegistrationInfo[serviceId]
+	if !foundRegistrationInfo {
+		return stacktrace.NewError("No registration info found for service '%v'", serviceId)
 	}
 	network.topology.RemoveService(serviceId)
-	delete(network.serviceIps, serviceId)
+	delete(network.serviceRegistrationInfo, serviceId)
 
 	// TODO PERF: Parallelize the shutdown of the service container and the sidecar container
-	serviceContainerId, foundContainerId := network.serviceContainerIds[serviceId]
-	if foundContainerId {
+
+	runInfo, foundRunInfo := network.serviceRunInfo[serviceId]
+	if foundRunInfo {
+		serviceContainerId := runInfo.containerId
 		// Make a best-effort attempt to stop the service container
 		logrus.Debugf("Stopping container ID '%v' for service ID '%v'...", serviceContainerId, serviceId)
 		if err := network.dockerManager.StopContainer(ctx, serviceContainerId, containerStopTimeout); err != nil {
 			return stacktrace.Propagate(err, "An error occurred stopping the container with ID %v", serviceContainerId)
 		}
-		delete(network.serviceContainerIds, serviceId)
+		delete(network.serviceRunInfo, serviceId)
 		logrus.Debugf("Successfully stopped container ID")
 	}
-	network.freeIpAddrTracker.ReleaseIpAddr(serviceIp)
+	network.freeIpAddrTracker.ReleaseIpAddr(registrationInfo.ipAddr)
 
 	sidecar, foundSidecar := network.networkingSidecars[serviceId]
 	if network.isPartitioningEnabled && foundSidecar {
@@ -477,21 +585,21 @@ NOTE: This is not thread-safe, so it must be within a function that locks mutex!
 func updateIpTables(
 		ctx context.Context,
 		targetBlocklists map[service_network_types.ServiceID]*service_network_types.ServiceIDSet,
-		serviceIps map[service_network_types.ServiceID]net.IP,
+		serviceRegistrationInfo map[service_network_types.ServiceID]serviceRegistrationInfo,
 		networkingSidecars map[service_network_types.ServiceID]networking_sidecar.NetworkingSidecar) error {
 	// TODO PERF: Run the container updates in parallel, with the container being modified being the most important
 	for serviceId, newBlocklist := range targetBlocklists {
 		allIpsToBlock := []net.IP{}
 		for _, serviceIdToBlock := range newBlocklist.Elems() {
-			ipToBlock, found := serviceIps[serviceIdToBlock]
+			infoForService, found := serviceRegistrationInfo[serviceIdToBlock]
 			if !found {
 				return stacktrace.NewError(
 					"Service with ID '%v' needs to block service with ID '%v', but the latter " +
-						"doesn't have an IP associated with it",
+						"doesn't have service registration info (i.e. an IP) associated with it",
 					serviceId,
 					serviceIdToBlock)
 			}
-			allIpsToBlock = append(allIpsToBlock, ipToBlock)
+			allIpsToBlock = append(allIpsToBlock, infoForService.ipAddr)
 		}
 
 		sidecar, found := networkingSidecars[serviceId]